move dynamic storage access into separate struct
